+++ examples/next-turbo-basic/next.config.mjs
@@ -1,0 +1,8 @@
+// @ts-check
+/** @type {import("next").NextConfig} */
+const nextConfig = {
+  reactStrictMode: true,
+  serverExternalPackages: ["esbuild", "@esbuild/linux-x64"],
+};
+
+export default nextConfig;