+++ examples/next-turbo-basic/pages/pages/index.tsx
@@ -1,0 +1,12 @@
+import Head from "next/head";
+
+export default function Home() {
+  return (
+    <>
+      <Head>
+        <title>Serwist example</title>
+      </Head>
+      <h1>Next.js + PWA = AWESOME!</h1>
+    </>
+  );
+}