+++ packages/core/src/utils/logger.ts
@@ -24,7 +24,7 @@ declare global {
  * before using it.
  */
 export const logger =
-  process.env.NODE_ENV === "production"
+  process.env.NODE_ENV === "production" || typeof self === "undefined"
     ? null!
     : (() => {
         // Don't overwrite this value if it's already set.