+++ packages/webpack-plugin/src/lib/schema.ts
@@ -1,28 +1,30 @@
-import { basePartial, injectPartial as baseInjectPartial, optionalSwDestPartial } from "@serwist/build/schema";
+import { fn, basePartial, injectPartial as baseInjectPartial, optionalSwDestPartial } from "@serwist/build/schema";
 import { z } from "zod";
 
-export const webpackPartial = z
-  .object({
-    chunks: z.array(z.string()).optional(),
-    exclude: z
-      .array(z.union([z.string(), z.instanceof(RegExp), z.function(z.tuple([z.any()]), z.boolean())]))
-      .default([/\.map$/, /^manifest.*\.js$/]),
-    excludeChunks: z.array(z.string()).optional(),
-    include: z.array(z.union([z.string(), z.instanceof(RegExp), z.function(z.tuple([z.any()]), z.boolean())])).optional(),
-  })
-  .strict("Do not pass invalid properties to WebpackPartial!");
+const webpackConditionCallback = fn({
+  input: [z.any()],
+  output: z.boolean(),
+});
 
-export const injectPartial = z
-  .object({
-    compileSrc: z.boolean().default(true),
-    swDest: z.string().optional(),
-    webpackCompilationPlugins: z.array(z.any()).optional(),
-  })
-  .strict("Do not pass invalid properties to WebpackInjectManifestPartial!");
+const webpackCondition = z.union([z.string(), z.instanceof(RegExp), webpackConditionCallback]);
 
-export const injectManifestOptions = basePartial
-  .merge(webpackPartial)
-  .merge(baseInjectPartial)
-  .merge(optionalSwDestPartial)
-  .merge(injectPartial)
-  .strict("Do not pass invalid properties to WebpackInjectManifestOptions!");
+export const webpackPartial = z.strictObject({
+  chunks: z.array(z.string()).optional(),
+  exclude: z.array(webpackCondition).default([/\.map$/, /^manifest.*\.js$/]),
+  excludeChunks: z.array(z.string()).optional(),
+  include: z.array(webpackCondition).optional(),
+});
+
+export const injectPartial = z.strictObject({
+  compileSrc: z.boolean().default(true),
+  swDest: z.string().optional(),
+  webpackCompilationPlugins: z.array(z.any()).optional(),
+});
+
+export const injectManifestOptions = z.strictObject({
+  ...basePartial.shape,
+  ...webpackPartial.shape,
+  ...baseInjectPartial.shape,
+  ...optionalSwDestPartial.shape,
+  ...injectPartial.shape,
+});